Handle SEGFAULT on exit
When exiting the 'main case' application, the SEGFAULT took place. The
reason was that JobManager::mInterpreter field (of type
std::shared_ptr<Interpreter>) was initialized with the pointer to the
statically allocated object. As result on exit the application
attempted to free its memory, which naturally resulted in error.

There is no statically allocated Interpreter instance in the
application any more. The only instance allocated on the heap, and the
only references to it are shared_ptr objects.